BREAKING CHANGE - in process of refactoring types and enums to new module, the migrations needed to be re-built as they had references to old python paths. Though... now, as I'm thinking about this, I think I can edit the migrations directly to prevent this from being breaking... probably best given this is released, though, admittedly with limited uptake. Also made a few tweaks in furtherance of headless crawlers being able to dump docs into an OpenContracts installation. Also removed some unwanted Affero GPL headers (this is NOT Affero licensed)
