+++ config/graphql/mutations.py
@@ -12,7 +12,7 @@ from django.db.models import Q
 from django.utils import timezone
 from graphene.types.generic import GenericScalar
 from graphql_jwt.decorators import login_required, user_passes_test
-from graphql_relay import from_global_id
+from graphql_relay import from_global_id, to_global_id
 
 from config.graphql.base import DRFDeletion, DRFMutation
 from config.graphql.graphene_types import (
@@ -56,9 +56,9 @@ from opencontractserver.tasks.permissioning_tasks import (
     make_analysis_public_task,
     make_corpus_public_task,
 )
+from opencontractserver.types.enums import PermissionTypes
 from opencontractserver.users.models import UserExport
-from opencontractserver.utils.data_types import PermissionTypes
-from opencontractserver.utils.permissioning_utils import (
+from opencontractserver.utils.permissioning import (
     set_permissions_for_obj_to_user,
     user_has_permission_for_obj,
 )
@@ -1101,12 +1101,14 @@ class CreateLabelForLabelsetMutation(graphene.Mutation):
     ok = graphene.Boolean()
     message = graphene.String()
     obj = graphene.Field(AnnotationLabelType)
+    obj_id = graphene.ID()
 
     @login_required
     def mutate(root, info, labelset_id, text, description, color, icon, label_type):
 
         ok = False
         obj = None
+        obj_id = None
 
         try:
             labelset = LabelSet.objects.get(
@@ -1121,6 +1123,7 @@ class CreateLabelForLabelsetMutation(graphene.Mutation):
                 label_type=label_type,
                 creator=info.context.user,
             )
+            obj_id = to_global_id("AnnotationLabelType", obj.id)
             logger.debug("CreateLabelForLabelsetMutation - mutate / Created label", obj)
 
             set_permissions_for_obj_to_user(
@@ -1138,7 +1141,9 @@ class CreateLabelForLabelsetMutation(graphene.Mutation):
         except Exception as e:
             message = f"Failed to create label for labelset due to error: {e}"
 
-        return CreateLabelForLabelsetMutation(obj=obj, message=message, ok=ok)
+        return CreateLabelForLabelsetMutation(
+            obj=obj, obj_id=obj_id, message=message, ok=ok
+        )
 
 
 class StartCorpusAnalysisMutation(graphene.Mutation):
