+++ opencontractserver/types/dicts.py
@@ -1,46 +1,8 @@
-import enum
 from typing import Optional, Union
 
 from typing_extensions import NotRequired, TypedDict
 
-
-class LabelType(str, enum.Enum):
-    DOC_TYPE_LABEL = "DOC_TYPE_LABEL"
-    TOKEN_LABEL = "TOKEN_LABEL"
-    RELATIONSHIP_LABEL = "RELATIONSHIP_LABEL"
-
-
-class JobStatus(str, enum.Enum):
-    CREATED = "CREATED"
-    QUEUED = "QUEUED"
-    RUNNING = "RUNNING"
-    COMPLETED = "COMPLETED"
-    FAILED = "FAILED"
-
-    @classmethod
-    def choices(cls):
-        return [(key, key) for key in cls]
-
-
-class PermissionTypes(str, enum.Enum):
-    CREATE = "CREATE"
-    READ = "READ"
-    EDIT = "EDIT"
-    UPDATE = "UPDATE"
-    DELETE = "DELETE"
-    PERMISSION = "PERMISSION"
-    PUBLISH = "PUBLISH"
-    CRUD = "CRUD"
-    ALL = "ALL"
-
-
-class AnnotationLabelPythonType(TypedDict):
-    id: str
-    color: str
-    description: str
-    icon: str
-    text: str
-    label_type: LabelType
+from opencontractserver.types.enums import AnnotationLabelPythonType
 
 
 class LabelLookupPythonType(TypedDict):
