+++ requirements/local.txt
@@ -27,8 +27,8 @@ sphinx-autobuild==2021.3.14 # https://github.com/GaretJax/sphinx-autobuild
 
 # Code quality
 # ------------------------------------------------------------------------------
-flake8==5.0.4  # https://github.com/PyCQA/flake8
-flake8-isort==5.0.0  # https://github.com/gforcada/flake8-isort
+flake8==6.0.0 # https://github.com/PyCQA/flake8
+flake8-isort==6.0.0  # https://github.com/gforcada/flake8-isort
 coverage==6.5.0  # https://github.com/nedbat/coveragepy
 black==22.10.0  # https://github.com/psf/black
 pylint-django==2.5.3  # https://github.com/PyCQA/pylint-django
